+++ src/us/mn/state/dot/tms/server/comm/DevicePoller.java
@@ -1,6 +1,6 @@
 /*
  * IRIS -- Intelligent Roadway Information System
- * Copyright (C) 2014-2020  Minnesota Department of Transportation
+ * Copyright (C) 2014-2022  Minnesota Department of Transportation
  * Copyright (C) 2017       SRF Consulting Group
  *
  * This program is free software; you can redistribute it and/or modify
@@ -24,10 +24,6 @@ import us.mn.state.dot.tms.server.ControllerImpl;
  * @author John L. Stanley
  */
 public interface DevicePoller {
-
-	/** Get the poller status.
-	 * Any value other than a blank string is considered "failed". */
-	String getStatus();
 
 	/** Check if the poller is currently connected */
 	boolean isConnected();