+++ src/us/mn/state/dot/tms/server/comm/ThreadedPoller.java
@@ -1,6 +1,6 @@
 /*
  * IRIS -- Intelligent Roadway Information System
- * Copyright (C) 2014-2020  Minnesota Department of Transportation
+ * Copyright (C) 2014-2022  Minnesota Department of Transportation
  * Copyright (C) 2015-2017  SRF Consulting Group
  *
  * This program is free software; you can redistribute it and/or modify
@@ -141,17 +141,10 @@ public class ThreadedPoller<T extends ControllerProperty>
 	/** Comm thread (may be null) */
 	private CommThread c_thread;
 
-	/** Get the poller status.
-	 * Any value other than a blank string is considered "failed". */
-	@Override
-	public synchronized String getStatus() {
-		return (c_thread != null) ? c_thread.getStatus() : "";
-	}
-
 	/** Check if the poller is currently connected */
 	@Override
 	public synchronized boolean isConnected() {
-		return (c_thread != null) && !c_thread.isDone();
+		return (c_thread != null) && c_thread.isConnected();
 	}
 
 	/** Get max seconds an idle connection should be left open